feat: add functionality to change style of Text of result. (#23)
* feat: add functionality to change style of Text of result.
Credits: https://github.com/fluttercommunity/flutter_google_places/commit/0041b20d8535f3bf31ff7a957b07917b931dc716

* fix
