+++ example/lib/main.dart
@@ -4,7 +4,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_google_places_hoc081098/flutter_google_places_hoc081098.dart';
 import 'package:google_api_headers/google_api_headers.dart';
 import 'package:google_maps_webservice/places.dart';
-import 'package:uuid/uuid.dart';
+
+import 'custom.dart';
 
 const kGoogleApiKey = 'API_KEY';
 
@@ -123,6 +124,9 @@ class _MyAppState extends State<MyApp> {
       mode: _mode,
       language: 'fr',
       components: [Component(Component.country, 'fr')],
+      // TODO: Since we supports Flutter >= 2.8.0
+      // ignore: deprecated_member_use
+      resultTextStyle: Theme.of(context).textTheme.subtitle1,
     );
 
     await displayPrediction(p, ScaffoldMessenger.of(context));
@@ -151,63 +155,4 @@ Future<void> displayPrediction(
       content: Text('${p.description} - $lat/$lng'),
     ),
   );
-}
-
-// custom scaffold that handle search
-// basically your widget need to extends [GooglePlacesAutocompleteWidget]
-// and your state [GooglePlacesAutocompleteState]
-class CustomSearchScaffold extends PlacesAutocompleteWidget {
-  CustomSearchScaffold({Key? key})
-      : super(
-          key: key,
-          apiKey: kGoogleApiKey,
-          sessionToken: const Uuid().v4(),
-          language: 'en',
-          components: [Component(Component.country, 'uk')],
-        );
-
-  @override
-  _CustomSearchScaffoldState createState() => _CustomSearchScaffoldState();
-}
-
-class _CustomSearchScaffoldState extends PlacesAutocompleteState {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const AppBarPlacesAutoCompleteTextField(
-          textStyle: null,
-          textDecoration: null,
-          cursorColor: null,
-        ),
-      ),
-      body: PlacesAutocompleteResult(
-        onTap: (p) => displayPrediction(p, ScaffoldMessenger.of(context)),
-        logo: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: const [FlutterLogo()],
-        ),
-      ),
-    );
-  }
-
-  @override
-  void onResponseError(PlacesAutocompleteResponse response) {
-    super.onResponseError(response);
-
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(content: Text(response.errorMessage ?? 'Unknown error')),
-    );
-  }
-
-  @override
-  void onResponse(PlacesAutocompleteResponse response) {
-    super.onResponse(response);
-
-    if (response.predictions.isNotEmpty) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text('Got answer')),
-      );
-    }
-  }
 }
